refactor: move segment to separate class
