+++ synthesis/filter/filter_segments.py
@@ -2,10 +2,7 @@ import numpy as np
 import json
 import ntpath
 import os
-from scipy.io import wavfile
-import parselmouth
-#from utilsfilter import *
-from g2p_en import G2p
+from .segment import Segment
 
 '''
 ## Features of interest
@@ -38,52 +35,6 @@ from g2p_en import G2p
 List of `Segments` representing valid segments
 '''
 
-# Uses the parselmouth library for audio data usage
-# https://parselmouth.readthedocs.io
-
-class Segment:
-    # todo extract parts
-    def __init__(self, path, start_time, end_time, samp_freq = 44100, text=''):
-        self.path = path # episode path
-        self.start_time = start_time
-        self.end_time = end_time
-        self.time = end_time - start_time
-        snd = parselmouth.Sound(path)
-        self.snd = snd.extract_part(
-            from_time = start_time,
-            to_time = end_time
-        )
-        self.data = self.snd.values.T # amplitudes
-        #print("LEN",len(self.data))
-
-        self.samp_freq = samp_freq
-        #self.spectrogram = self.snd.to_spectrogram()
-        self.intensity = self.snd.to_intensity()
-        self.pitch_obj = self.snd.to_pitch()
-        self.pitch = self.pitch_obj.selected_array['frequency']
-        self.energy = self.snd.get_energy()
-
-        self.text = text
-
-    def write(self,path): 
-       wavfile.write(path, self.samp_freq, self.data) 
-
-    def pitch_avg(self):
-        return np.average(self.pitch)
-
-    def get_energy(self):
-        return self.energy
-
-    def intensity_avg(self):
-        return np.average(self.intensity)
-    
-    def get_speech_rate(self):
-        # Uses g2pE for phoneme conversion
-        # https://github.com/Kyubyong/g2p
-        g2p = G2p()
-        phonemes = g2p(self.text)
-        phonemes = [s for s in phonemes if s.strip()]
-        return len(phonemes) / self.time
 
 
 def arg_parse(args):
